feat: add functionality to show popup and set custom price for selected binding type
The changes in this commit add functionality to show a popup and set a custom price for the selected binding type. 

Previously, when fetching the binding cost data from the backend, if the data was not available, the binding cost state was not updated and the popup was not shown. This behavior has been changed. Now, if the binding cost data is not available, the `setIsLaminationSelected` state is set to true and the `setShowPopup` state is set to true, indicating that a custom price needs to be entered.

Similarly, for the lamination type, if the selected lamination type requires a custom price, the `setIsLaminationSelected` state is set to true and the `setShowPopup` state is set to true. If the lamination cost data is available, the `setLaminationPrice` state is updated with the fetched lamination cost value.