style: update colors and fonts in header bar and buttons
The changes were made to update the colors and fonts in the header bar and buttons. The color of the header bar text was changed to #404146, and the background color of the buttons was changed to #1f4e5f and #6d9d88 respectively. The font family for the header bar and buttons was also updated to "DM Sans" and "Rubik" respectively.